bundles/core/src/test/java/io/wcm/wcm/core/components/impl/models/helpers/LinkWrapperTest.java
/*
 * #%L
 * wcm.io
 * %%
 * Copyright (C) 2019 wcm.io
 * %%
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *      http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 * #L%
 */
package io.wcm.wcm.core.components.impl.models.helpers;

import static io.wcm.samples.core.testcontext.AppAemContext.CONTENT_ROOT;
import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.junit.jupiter.api.Assertions.assertFalse;
import static org.junit.jupiter.api.Assertions.assertNull;
import static org.junit.jupiter.api.Assertions.assertTrue;

import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;

import com.day.cq.wcm.api.Page;

import io.wcm.handler.link.Link;
import io.wcm.handler.link.LinkHandler;
import io.wcm.samples.core.testcontext.AppAemContext;
import io.wcm.sling.commons.adapter.AdaptTo;
import io.wcm.testing.mock.aem.junit5.AemContext;
import io.wcm.testing.mock.aem.junit5.AemContextExtension;

@ExtendWith(AemContextExtension.class)
class LinkWrapperTest {

  private final AemContext context = AppAemContext.newAemContext();

  private LinkHandler linkHandler;

  @BeforeEach
  void setUp() {
    linkHandler = AdaptTo.notNull(context.request(), LinkHandler.class);
  }

  @Test
  void testValidLink() {
    Page page = context.pageManager().getPage(CONTENT_ROOT);
    Link link = linkHandler.get(page).build();
    LinkWrapper underTest = new LinkWrapper(link);

    assertTrue(underTest.isValid());
    assertEquals(link.getUrl(), underTest.getURL());
    assertEquals(link.getAnchorAttributes(), underTest.getHtmlAttributes());
    assertEquals(link.getTargetPage(), underTest.getTargetPage());
  }

  @Test
  void testInvalidLink() {
    Link link = linkHandler.invalid();
    LinkWrapper underTest = new LinkWrapper(link);

    assertFalse(underTest.isValid());
    assertNull(underTest.getURL());
    assertNull(underTest.getHtmlAttributes());
    assertNull(underTest.getTargetPage());
  }

}
